+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove bogus error logs on subsequent logins due to tutorial guidepoints already having been added on first login.
+
 ## [0.6.28] - 2024-06-24
 
 ### Removed